+++ DX11 Game/Graphics/Viewport.h
@@ -13,7 +13,7 @@ namespace Bind
 		Viewport( Graphics& gfx ) : Viewport( gfx, gfx.GetWidth(), gfx.GetHeight() ) {}
 		Viewport( Graphics& gfx, UINT width, UINT height )
 		{
-			viewportDesc = CD3D11_VIEWPORT( 0.0f, 0.0f, width, height );
+			viewportDesc = CD3D11_VIEWPORT( 0.0f, 0.0f, static_cast<FLOAT>( width ), static_cast<FLOAT>( height ) );
 			GetContext( gfx )->RSSetViewports( 1u, &viewportDesc );
 		}
 	private:
